requirements.txt
PyQt5==5.15.6
python-dotenv==0.20.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
